Export to grey (#29)
* Add severity levels to ValidationReporter

* Re-design ddln.yaml file write

* Add CsvDirectoryExporter.clear_if_exists parameter

* Implement grey-export endpoint

* Add grey export to front-end

* Calculate ddln_id on export to grey

* Fix nfs permissions issue

* Export to grey with right permissions

* Run annotation export in worker process

* Copy logs from previous merge attempt

* Make sequence checkboxes more convenient to use

* Download merge result on button click, rather than automatically

* Fix download merge result not working after request extra anno is piressed

* Fix a typo

* Implement accept segments endpoint

* Update export grey endpoint to handle multiannotations

* Add export button for multiannotation task

* Hide GreyExporter component for multiannotation task

* Add accept sequences button

* Use PyYAML to read/write yaml files

* Update invalid.yaml file on manual sequence acceptance

* Change track_id_non_consecutive check's severity to WARNING

* Make team section simpler in ddln.yaml file

Change it to dictionary rather than array of dictionaries

* Address code review comments

+++ cvat/apps/annotation/transports/csv/exporter.py
@@ -6,9 +6,9 @@ from pathlib import Path
 
 
 class CsvDirectoryExporter:
-    def __init__(self, base_dir):
+    def __init__(self, base_dir, clear_if_exists=True):
         self._base_dir_path = Path(base_dir)
-        if self._base_dir_path.exists():
+        if clear_if_exists and self._base_dir_path.exists():
             shutil.rmtree(str(self._base_dir_path))
 
     def __enter__(self):
